Move `OpErrors` to separate module
